pico: split audio drivers
Also, drop i2s from the default config while we're here.

This makes the PicoSystem PWM override generic (-DBLIT_AUDIO_DRIVER=pwm instead of -DAUDIO_PWM).
There is no checking if the driver is valid though. (none is always valid)
