+++ 32blit-pico/audio_none.cpp
@@ -1,0 +1,7 @@
+#include "audio.hpp"
+
+void init_audio() {
+}
+
+void update_audio(uint32_t time) {
+}
