+++ 32blit-pico/board/pimoroni_picosystem/config.cmake
@@ -1,25 +1,17 @@
 set(BLIT_BOARD_NAME "PicoSystem")
 
 set(BLIT_BOARD_DEFINITIONS
+    AUDIO_BEEP_PIN=PICOSYSTEM_AUDIO_PIN
     DISPLAY_ST7789
     INPUT_GPIO
     LED_R_PIN=PICOSYSTEM_LED_R_PIN
     LED_G_PIN=PICOSYSTEM_LED_G_PIN
     LED_B_PIN=PICOSYSTEM_LED_B_PIN
     USB_PRODUCT_STR="PicoSystem"
+
+    # non-default PWM audio support
+    PICO_AUDIO_PWM_MONO_PIN=PICOSYSTEM_AUDIO_PIN
+    PICO_AUDIO_PWM_PIO=1
 )
 
-if(AUDIO_PWM)
-    set(BLIT_BOARD_LIBRARIES pico_audio_pwm)
-    list(APPEND BLIT_BOARD_DEFINITIONS
-        PICO_AUDIO_PWM_MONO_PIN=PICOSYSTEM_AUDIO_PIN
-        PICO_AUDIO_PWM_PIO=1
-        AUDIO_PWM
-    )
-    set(BLIT_REQUIRE_PICO_EXTRAS TRUE) # audio_pwm
-else()
-    list(APPEND BLIT_BOARD_DEFINITIONS
-        AUDIO_BEEP_PIN=PICOSYSTEM_AUDIO_PIN
-        AUDIO_BEEP
-    )
-endif()
+blit_driver(audio beep)
