+++ 32blit-pico/board/vgaboard/config.cmake
@@ -3,10 +3,11 @@ set(BLIT_BOARD_NAME "VGA Board")
 set(BLIT_BOARD_DEFINITIONS
     PICO_SCANVIDEO_PLANE1_VARIABLE_FRAGMENT_DMA=1
     PICO_SCANVIDEO_MAX_SCANLINE_BUFFER_WORDS=12
-    AUDIO_I2S
     DISPLAY_SCANVIDEO
 )
 
-set(BLIT_BOARD_LIBRARIES pico_scanvideo_dpi pico_audio_i2s)
+set(BLIT_BOARD_LIBRARIES pico_scanvideo_dpi)
 
-set(BLIT_REQUIRE_PICO_EXTRAS TRUE) # scanvideo/audio_i2s
+set(BLIT_REQUIRE_PICO_EXTRAS TRUE) # scanvideo
+
+blit_driver(audio i2s)
